// src/features/bloggers-platform/blogs/api/models/output/blog.output.model.ts
import { SortQueryFilterType } from '../../../../../../base/pagination.base.model';
import { SearchNameTermFilterType } from '../../../../../../base/pagination.base.model';

export class BlogOutputModel {
  id: string;
  name: string;
  description: string;
  websiteUrl: string;
  createdAt: Date;
  isMembership: boolean;
}

export interface QueryBlogFilterType
  extends SortQueryFilterType,
    SearchNameTermFilterType {}